hardware/raspberry-pi/requirements.txt
paho-mqtt==2.1.0
python-dotenv==1.0.1
smbus3==0.5.5
